refactor: folder structure and routes structure
