--- front-end/src/components/PrivateRoute.tsx
@@ -1,26 +1,0 @@
-import { ReactNode } from "react";
-import { Navigate } from "react-router-dom";
-import { useAuth } from "../lib/AuthContext";
-import Layout from "./Layout";
-
-interface PrivateRouteProps {
-  children: ReactNode;
-}
-
-export default function PrivateRoute({ children }: PrivateRouteProps) {
-  const { currentUser, loading } = useAuth();
-
-  if (loading) {
-    return (
-      <div className="flex justify-center items-center h-[50vh]">
-        Carregando...
-      </div>
-    );
-  }
-
-  if (!currentUser) {
-    return <Navigate to="/login" />;
-  }
-
-  return <Layout>{children}</Layout>;
-}
