+++ front-end/src/components/ProtectedLayout.tsx
@@ -1,0 +1,12 @@
+import { Navigate, Outlet } from "react-router-dom";
+import { useAuth } from "../lib/AuthContext";
+
+export default function ProtectedLayout() {
+  const { currentUser } = useAuth();
+
+  if (!currentUser) {
+    return <Navigate to="/login" />;
+  }
+
+  return <Outlet />;
+}
