+++ front-end/src/pages/public/LandingPage.tsx
