+++ front-end/src/routes.tsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from "react-router-dom";
-import Login from "./pages/Login";
-import Dashboard from "./pages/Dashboard";
-import PrivateRoute from "./components/PrivateRoute";
-import LandingPage from "./pages/LandingPage";
+import LandingPage from "./pages/public/LandingPage";
+import Dashboard from "./pages/private/Dashboard";
+import Login from "./pages/public/Login";
+import ProtectedLayout from "./components/ProtectedLayout";
 
 export const router = createBrowserRouter([
   {
@@ -16,13 +16,15 @@ export const router = createBrowserRouter([
         path: "login",
         element: <Login />,
       },
+    ],
+  },
+  {
+    path: "/panel",
+    element: <ProtectedLayout />,
+    children: [
       {
         path: "dashboard",
-        element: (
-          <PrivateRoute>
-            <Dashboard />
-          </PrivateRoute>
-        ),
+        element: <Dashboard />,
       },
     ],
   },
